[Findit] Use target master/builder property when indexing builds.
This change makes the ingestor use the target mastername and buildername
for indexing the isolated artifacts when those properties are present.

Without this change, the builder that produces the build is indexed but
the builder may not represent the configuration used for the build. Such
as in the case of luci.chromium.findit/findit_variable, which we use for
all tryjob builds.

R=chanli,stgao

Change-Id: I21c0110db6ee7b0ba3d12dfad66c1eca50a4d92a
Reviewed-on: https://chromium-review.googlesource.com/1114205
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>