[third_party_packages/cmake] fix Mac bootstrap
Commit 4113551a37776b702187b31642f88d5162e966ae allows `cmake` to be
built using a bootstrap; however, the bootstrap version that the recipe
uses, "3.11.4", doesn't exist for "mac-amd64", causing the build to
fail.

This is a temporary band-aid until the same bootstrap version exists on
all platforms. It's just a bootstrap, though, so probably not urgent.

Change-Id: I8bdc452b054c82f631022847b66221ad7cbb31f1
Reviewed-on: https://chromium-review.googlesource.com/1115967
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>