crosskylabadmin: Move swarming client to separate package.
The swarming client adaptor is now moved to a clients/ package. This
package is not named swarming because users need to reference the
generated swarming bindings along with this package. Any client side
adaptors that allow using stubs for unittests should live in this
package.

BUG=chromium:846020

Change-Id: Ie3a56e5597beb77eec9b4d62bb331e96bf30a33c
Reviewed-on: https://chromium-review.googlesource.com/1116063
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>