[Findit] Use full change id when query Gerrit.
If we query Gerrit with the change id of a CL that is merged into a different
branch after it landed in the master branch, Gerrit responses with 404 because
the query result includes multiple CLs on Gerrit.

Using commit:git_hash to query on Gerrit does return a unique CL even it is
merged into a different branch other than master.

Change-Id: I25c4472c99c9b7915af46ccd88e202da391129f2
Reviewed-on: https://chromium-review.googlesource.com/1118914
Commit-Queue: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>