[tricium] Do not consider it an error when there is no isolated output
In general, in the logs we are seeing two pubsub notifications
per worker for Tricium functions, the first one is received before
the task is finished, and the second one afterwards.

This changes the behavior of Tricium to do nothing when trying to
collect and finding that there is no isolated output hash yet, so
that it will just do nothing and return 200 response.
The output should still be collected after the second pubsub
notification comes along.

Note: There's a potential concern though that
if swarming's behavior changes so there's not always a pubsub
notification after the task is done, some task output will stay
uncollected and analyzer runs may stay unfinished.

Bug: 803996
Change-Id: I253c777c6f8328bf307deee90cf7f2b576f1a327
Reviewed-on: https://chromium-review.googlesource.com/1123062
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>