Unmarshal comments with jsonpb when reporting results
In https://crrev.com/c/1121425 I changed some JSON serialization
from the standard json library to jsonpb, but it was still being
decoded using the standard json library.

What this means is that line was being encoded as startLine,
but when decoding it was looking for start_line, and finding
nothing.

Bug: 853887
Change-Id: I5054f7bfad74723b0f2749dd34ecd3e05debf4e9
Reviewed-on: https://chromium-review.googlesource.com/1123647
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>