[Tricium] Also use jsonpb also when unmarshaling comments in gerrit.go
This is the 3rd in a series of CLs changing JSON marshaling:

 - First: https://crrev.com/c/1121425
 - Second: https://crrev.com/c/1123647

After the second CL but before this CL, line numbers are stored in
track.Comment but not reported to Gerrit; the reason (again) was
because the standard json marshaler was used to unmarshal comments
before sending them. This CL changes that.

This CL also includes minor variable name changes for the purpose
of improving consistency and clarity.

Bug: 853887
Change-Id: I3337b03a81316842b1b9c3d6af95fd14d3e87cdc
Reviewed-on: https://chromium-review.googlesource.com/1127351
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>