[tricium] Change the collect request to succeed but reenqueue
This is a re-commit https://crrev.com/c/1123062 plus an
extra change to reenqueue the collect task when there
is no output yet.

Previously, when a lack of isolated output meant the task
was considered failed, it was retried until it was successful.

It appeared that there were two pubsub notify requests so
ignoring the first would be OK, but after deploying to
tricium-dev I saw that actually it seems to result in tasks
that never get their output collected and never get marked
complete.

This CL would make it so that when a collect request is
made but finds no isolated output, it retries repeatedly
until it succeeds.

Bug: 1123062
Change-Id: I4dd2525a0ddbae4ab71954bb35b42989fea5c1df
Reviewed-on: https://chromium-review.googlesource.com/1125303
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>