[Monorail] Implement a basic bar chart component
This introduces an <mr-bar-chart> element for basic bar charts. The
chart in this CL is super basic and has no controls. Once this is landed
more functionality will come in follow-up CLs.

Demo:
https://16438-d3a47d7-tainted-jeffcarp-dot-monorail-staging.appspot.com/p/chromium/issues/list?can=2&mode=chart

The demo also interprets the &q= parameter. In order to change it you'll
need to either submit a new query and manually set the query param back to
&mode=chart, or edit the &q= param directly and reload.

Bug: monorail:3339
Change-Id: Ie099417fdc16d749018161e07f13d322c305fc1c
Reviewed-on: https://chromium-review.googlesource.com/1125330
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>