[tricium] Omit empty "id" field in robotCommentInput JSON
Non-functional, minor change.

The id field exists in RobotCommentInput since it exists in CommentInfo,
but it's not actually meant to be included in the request since
it's ignored and generated by Gerrit.

https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#comment-info

This just adds a omitempty annotation to the struct; this CL would also
changes the JSON-serialized string in the test to include indentation.

Change-Id: Ied1c31a624d86400c9b91a95a5ba552553c4f007
Reviewed-on: https://chromium-review.googlesource.com/1137214
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>