[Findit] Don't store log_data in step if it's too large.
This change will properly handles the case that log_data of a step is too large and cannot be stored in data store. Although after this change, we may need to request isolate for the failure_logs again when to use the log.

I also refactored functions in ci_test_failure a bit to make it easier to understand and use.

Also removed some data file for unittests.

Bug: 806406
Change-Id: I570cd3dfa9a30809764e80e22e60504ff2a2a482
Reviewed-on: https://chromium-review.googlesource.com/1100545
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>