[tricium] Manage configs explicitly with a cron job and datastore
Background: Currently Tricium uses the
go.chromium.org/luci/config/server/cfgclient library
for fetching configs when they're needed. This appears to
have two disadvantages:

1. Ocasionally when a config is actually fetched, it takes
   about 5 seconds, occasionally adding delay to other
   operations (polling, starting analysis, etc.)
2. Sometimes a new config is not fetched when it should be,
   and a stale cached version is used (crbug.com/852001).

This CL changes it so that configs are explicitly stored and
managed in the Tricium datastore, and updated periodically
by a cron job.

This CL keeps the interface and mocks the same, so that
other code that uses configs doesn't have to care about
how the configs are fetched or managed.

Bug: 852001

Change-Id: I6a8740ce2edfb5691a6b98729a1611b0b4985d28
Reviewed-on: https://chromium-review.googlesource.com/1132547
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>