[tricium] Store owner/component in tracked FunctionRun
In this CL:
 - Add function definitions to generated Workflow in config.Generate
 - Add owner/component to stored FunctionRun entity in WorkflowLaunched handler.

In the Launch RPC, we're already fetching the project and service
configs and getting the function definitions that will be used in
the current run; we can store the owner/component at this point
to avoid having to fetch these things again later on.

If we didn't store things here, then later on when these things are needed,
the configs would have to be fetched & merged later on in order to get the
analyzer owner and component.

Proposed next CL: Add owner/component to ReportNotUsefulReponse;
it would also be possible to put owner/componeont in each commment.

One-pager doc related to this change:
https://docs.google.com/document/d/1-TRt0Wwm-6-eh92eFrQJsYDp_awAuizOZGjPfmrWkkY/edit

Bug: 865200
Change-Id: Ie35a77ecfce4bfb04044b3948719785f0e357d02
Reviewed-on: https://chromium-review.googlesource.com/1147213
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>