UserCommits Cron
Added entire cron. If the usercommits table is over 20k, the oldest columns will be dropped to make it 20k. When a new repo is added, no more than 10k commits are added, and no commit older than the oldest commit in the usercommits table is added.

Bug: monorail:
Change-Id: Iecdf430fe22e65841b7e66f9c9d59d45954c21d6
Reviewed-on: https://chromium-review.googlesource.com/1137743
Commit-Queue: Markie Wagner <markie@google.com>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>