[Findit] Flake Analyzer - Adding assert for unsupported try jobs.
Several flake analyses fail when trying to run try jobs. The root cause
is there is no trybot configured for the given master/builder, however
this reason is not clear. This change serves only to add an informative
assert message to make determining this root cause easier.

The real fix is to add the corresponding trybot.

Change-Id: Ie97d1c35793b0b5e62ec11e98e55fbb1386f24ae
Reviewed-on: https://chromium-review.googlesource.com/1150912
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>