[tricium] Refactor pubsub.go, use luci common package
This is meant to be a small refactoring; this changes the logic used
to get topic/subscription strings to use luci/common/gcloud/pubsub
so that this module doesn't have to contain knowledge about how
such strings are constructed.

The original purpose of this was to resolve a TODO left by Emma.

Change-Id: I9bbe8ad9d449dbf1586cf33613f7fc3c33331a6e
Reviewed-on: https://chromium-review.googlesource.com/1153843
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>