[tricium] Validate function names.
If someone accidentally adds a function with an underscore in
the name, then it would be split incorrectly when extracting
name/platform from a worker name, since worker names are made from
function and platform names separated with an undersore.

This is a safeguard against someone accidentally adding such a
function name later.

Change-Id: Iec75f5e301815c3bb114607dc1d8b969ad906ff5
Reviewed-on: https://chromium-review.googlesource.com/1153658
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>