[tricium] Track workers that have no output
This changes the validation of worker done requests
so that requests with no output are accepted.

This means that some requests with no output will
go on to be processed, and if the worker is an analyzer,
we want to just report no comments and set the state
to whatever the request says to set the state to
(e.g. FAILURE if the return code is non-zero).

Bug: 868595
Change-Id: Iec0315e9e15329d03663df578c4f1149613297ef
Reviewed-on: https://chromium-review.googlesource.com/1157137
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>