[tricium GitFileIsolator] Follow-up fix and test update
Before submitting https://crrev.com/c//1159111, I didn't
run the isolator with the test input, which would have caught
the bug (trying to stat relative path, which was incorrect).

This CL fixes that and updates the test to a test case that
Vadim pointed out a while ago. After running this I see that
it behaves as expected, skipping symlinks and writing output
which doesn't include the symlinks.

Bug: 848817
Change-Id: I1919fb4ca85928082fc40d0fa997228ea748d9de
Reviewed-on: https://chromium-review.googlesource.com/1161017
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>