[Findit] Flake Analyzer - Fixing check for stability to use proper thresholds
This is part 1 of a 2-part fix. The root cause is the config values are being
read "incorrectly," with upper/lower_flake_threshold being under "swarming_rerun"
within flake settings. The code that reads these values reads them at the root
level (as it should), causing the values not to be found and thus using fallback
values defined in flake_constants.py which still use the old 0.02, 0.98 figure.

In Part 1 here, the fix is to use pass_rate_util.IsStableDefaultThresholds which
will always read from config and to update the fallback values to 1e-7 and
0.9999999.

Part 2 will be to clean up config so it is in the expected format, captured in
issue 799832.

Bug: 870376
Change-Id: Ie7ddd85d558b22c94b2991e8808bc7936115d28c
Reviewed-on: https://chromium-review.googlesource.com/1162665
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>