[tricium spellchecker] Add more file types to comment_formats.json
For now, this is based on a quick survey of the file types in Chromium:

find chromium/src | sed 's/.*\.//' | sort | uniq -c | sort -nr

   freq extension
  89200 h
  74490 html
  54348 txt
  46824 cc
  37755 png (binary)
  33320 js
  31171 wasm (binary)
  31149 o (binary)
  17282 py
  15148 cpp
   9693 c
   7115 json
   6208 so (binary)
   5527 stamp (binary?)
   5433 ll
   5094 java
   4709 hpp
   4685 xht
   4440 ninja
   4374 frag
   ...

Some of these are just in third_party and may not be
actively modified though, like .ll and .frag.
The reason for the huge number of .txt, .html and .png
is probably the layout tests.

Other file types will be present in other repos, e.g.
rust files in fuchsia. This list may have to be expanded
a bit more.

Bug: 872037
Change-Id: I2343d4c060af04284ab9d3a159439cb4d22cddbd
Reviewed-on: https://chromium-review.googlesource.com/1167890
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>