[Findit] Flake Analyzer - Analysis status not necessary for auto revert
Analysis.status is now only set at the very end of AnalyzeFlakePipeline
upon completion of OnFinalized at the root level, which is after the
logic for auto revert has ran. It is unnecessary for auto revert to check
the analysis's running status, as it is possible to revert once there's
a culprit with sufficient confidence.

Bug: 872838
Change-Id: I8ce0c789f07a69c89b22e2b4c5840c6c5b915bee
Reviewed-on: https://chromium-review.googlesource.com/1169863
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>