[luci-migration] Track LUCI-is-prod builders and stop marking bugs migrated.
Buildbot builders can sometimes fail to be discovered by the migration app.

If we "lost" a Buildbot builder and there is no data from the LUCI builder, it
may have been deleted or renamed, or there was a blip in liveness signal.
  * In the former case, we just want to delete the builder entity.
  * In the latter case, we will restore the builder entity when it signals life
    again with no LUCI data loss.

If there is data from the LUCI builder, we want to retain the builder entity so
there is a record of the builder, and not mark the bug as fixed (in case the
builder returns).

Also, surface number and percentage of LUCI-is-prod builders on the app because
true migratiion status encompasses decommission state, which is known only to
the bug, and not to the app.

Change-Id: Iadf1e91d1754d5ccd2ae4570cb0a0555bdd408fb
Reviewed-on: https://chromium-review.googlesource.com/1159860
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>