[Findit] Add function to get isolate_target_name
This is a pre-requirement CL for mapping different step names to the
same base test target that actually define the tests. For example:
viz_browser_tests -> browser_tests.
site_per_process_webkit_layout_tests -> webkit_layout_tests.

The utlimate solution is to add a base_test_target field in the GN
templates that defines the tests, but unfortunately, it doesn't exist
yet.

As of now, the closest thing we have is the isolate_target_name in the
step_metadata. Though there is a refactoring going on by Dirk that may
change the values of the isolate_target_name, I've synced with him and
he promised that he will make sure isolate_target_name stays the same
until the base_test_target is in place.

So, the conclusion is that it is safe to use isolate_target_name as a
temporary solution.

Bug: 848867
Change-Id: I73f36c8a8ef7f3bcf0db3869bf0bc89076f2a421
Reviewed-on: https://chromium-review.googlesource.com/1171248
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>