[Findit] Flake Analyzer - Implementing fallback logic when mapping commit position -> nearby builds in NextCommitPositionPipeline
Not all builders are on LUCI yet, and thus flake analysis would fail when
trying to map a commit position to an IsolatedTarget. In order to get build
shas to trigger swarming tasks against for all builds supported for
analysis, fallback to searching buildbot for nearby builds when IsolatedTargets
are unavailable.

This is a temporary workaround until all builds have been migrated to
LUCI, and this CL can be reverted or an inverse manually created.

Bug: 872515
Change-Id: Id1063b158b716f47cb8b3257162fc2191bf1c827
Reviewed-on: https://chromium-review.googlesource.com/1170358
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>