[tricium spellchecker] Rearrange unit test methods
This is a minor refactoring that is meant to make
it slightly easier to add future tests. This CL adds
a simpler (one-word) example case, and also updates
test descriptions and order.

Also, I think I noticed a bug in the suggestion range,
so a TODO is added for that.

Change-Id: Ic4c0ba60da3636658f581d0d075fbc47093a72e5
Reviewed-on: https://chromium-review.googlesource.com/1173054
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>