[Findit] Make Flake Analyzer check open bugs for labels and test names
In the ShouldFileBugForAnalysis function, it returns False when either
BugAlreadyExistsForCustomField or OpenBugAlreadyExistsForTest is True
because Flake Analyzer thinks bugs already exists, and it should update
them instead of creating new ones.

However, the create bug pipeline doesn't attach the bug id to the
analysis object, which makes the update bug pipeline not able to find
the correct bugs to update.

Change-Id: I56a48f256549669cde83db4983c973c283fe39ae
Reviewed-on: https://chromium-review.googlesource.com/1176178
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>