[FindIt] Remove unncessary APIs in issue_tracking_service
For each public API to search for a bug and return its id, there is
a counterpart that does the exact same thing, but returns the and issue
object instead.

One obvious problem is that they have very similar names that is quite
hard to tel them apart, which I think is error-prone. After checking
all the callers, it seems that all of them only need the id, so I want
to remove the ones that return issue objects in this CL.

Change-Id: I5d254567e8cef7c2f028d452cf12b5b486716c29
Reviewed-on: https://chromium-review.googlesource.com/1176598
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>