qscheduler: normalize protos
Prior to this change, a number of proto types in this project had a
field specifying their own ID, but were also stored by parent objects in
an ID-to-instance map. This is redundant, and is a potential source of
inconsistent data, and was leading to extra meaningless boilerplate in
test code. Eliminate this redundancy in the schema.

BUG=chromium:870537
TEST=unit test

Change-Id: I650559249aa53ed793b156d6087c5caaca9614b3
Reviewed-on: https://chromium-review.googlesource.com/1175556
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>