qscheduler: add a tutils utility package for converting time protos
The standard ptypes library returns an error on proto parsing errors
when converting to/from timestamp.Timestamp or duration.Duration. In
most cases, callers would find it more convenient to simply panic on
parsing errors, and allow calls to the conversion func to be inline with
timestamp assignments etc.

This implements a tiny convenience library for this purpose.

BUG=chromium:870537
TEST=unit test

Change-Id: I459879ce7e75139b483be4b09ff1c65d3df40fc3
Reviewed-on: https://chromium-review.googlesource.com/1180374
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Reviewed-by: smut <smut@google.com>