[Monorail] Perf: fix location of autocomplete tracking
I mistakenly put the logger.logEnd call at the end of the function that
sets up the autocomplete population instead of at the end of the callbacks in
that function, resulting in values of 0 showing up in GA for this timing metric.
This CL puts the logEnd call in the correct location.

Bug: monorail:4052
Change-Id: Ifdbfa808618fe1f22cea244a6eb6420c97f1774c
Reviewed-on: https://chromium-review.googlesource.com/1185789
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>