[FindIt] Implement issue generator and change Flake Detection to use it
This CL creates the IssueGenerator class, and the intention is to
provide a unified bug filing interface and logic that can be applied
to both Flake Detection and Flake Analyzer.

This CL first switches Flake Detection to use the new interface, and
once it sticks, will switch Flake Analyzer to use it.

For more details, please refer to:
https://docs.google.com/document/d/1CIvy_3pelb5fbzOsm4mc-Gg_-1pvivEx0WowsfzgF1M/edit#heading=h.lp7pod8di42k

Bug: 875899
Change-Id: I45fd75c53d9588c8bc668b9822a6227bf5f71658
Reviewed-on: https://chromium-review.googlesource.com/1182512
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>