[Monorail] Clarify some field-value code.
I just happened to look at this again while figuring out something
for Tiff.  We don't use the verb "eject" anywhere else in the code.
And, the possible field types are all mutually exclusive, so the
conditionals are more clear as an if-else chain.

R=ehmaldonado@chromium.org

Change-Id: Ibc220f3e91fe08ddb20a95f202b6c52a9bc5134f
Reviewed-on: https://chromium-review.googlesource.com/1191963
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>