quotascheduler: encapsulate scheduler logic
This change moves the bulk of scheduler logic (which was formerly spread
across several packages) into a single package, scheduler, and defines
an exported type Scheduler with methods to interact with the
quotascheduler algorithm. Details about the implementation of the
algorithm are no longer exported.

This CL does not introduce any behavioral changes. A few new methods
that Scheduler will implement are not yet defined, because it is easier
to first accomplish non-behavior-changing refactor in this CL, and
behavior changes in a separate follow-up.

BUG=chromium:870537
TEST=unit tests

Change-Id: I7017b6f287855006a9134aa0c4a159adb8684baa
Reviewed-on: https://chromium-review.googlesource.com/1189426
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>