Import skylab_swarming_worker Go command
This code was originally sharing a repo in Chromium OS:

https://chromium.googlesource.com/chromiumos/infra/lucifer/

The code is being moved to infra.git because it does not need to be built into a Chromium OS image
and it is easier to work on and test using the tools around infra.git than in the Chromium OS repo.

The code is imported as of https://chromium-review.googlesource.com/c/chromiumos/infra/lucifer/+/1137490
and minimal changes were made to the code to facilitate the import.
Import paths were changed to reflect the new location, and minimal changes were
made to pass lint and tests, and to accommodate different versions of dependencies.
The relative paths of some subpackages were changed because all of the import
paths had to be fixed anyway and to hide everything under internal
for the initial import.

Change-Id: I5a9366a843d49810f556841e3db30042cefb41af
Reviewed-on: https://chromium-review.googlesource.com/1187554
Commit-Queue: Allen Li <ayatane@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>