[Monorail] Get flipper working in Firefox and Edge by waiting on polyfill event
This works in FF62 and Edge 17, but it's potentially more hacky than it needs
to be.

I believe the root problem of https://crbug.com/monorail/4177 is that the
webcomponents polyfill is not loaded by the time mr-flipper is loaded. The
polyfill emits an event, WebComponentsReady, once loaded. Loading mr-flipper
after this event fires fixes the problem.

Demo:
https://17294-446866e-tainted-jeffcarp-dot-monorail-staging.appspot.com

Bug: monorail:4177
Change-Id: I7ae1c7be58a9f786bd40e8b317c76a7ecc9d0ab2
Reviewed-on: https://chromium-review.googlesource.com/1188537
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>