[Findit] Key IsolatedTarget by build_id/target/isolate_hash.
Instead of the previous isolate_hash-only key. This allows us to record
entries for builds that have the same isolated hash as older builds.
Which is expected due to incremental compilation.

This change replaces the isolated_hash property of the model to use the
new column when available and fall back to the key, so that it works
with the old records as well.

Also remove the .Get method which isn't used except for tests and
replace its use with .get_by_id included as part of ndb.Model.

R=stgao,chanli,lijeffrey,liaoyuke

Bug: 878552
Change-Id: Ib8dd3f5c42d513283152bc49aca22350e47fe0bc
Reviewed-on: https://chromium-review.googlesource.com/1194893
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>