[Monorail] Replace account dropdown with Polymer dropdown
It was brought to our attention that the dropdown menus in the Monorail
header are not accessible: https://crbug.com/monorail/4132. This replaces
the account header with a Polymer element. If this is successful, we'll
replace the other dropdown menus in the header with Polymer elements.
Eventually the FLT header and Monorail header should merge into one component.

Demo:
https://17317-3e40535-tainted-jeffcarp-dot-monorail-staging.appspot.com/

Bug: monorail:4132
Change-Id: I159445ee28bc088ad0fe5af9058d3ef2bbad4cc1
Reviewed-on: https://chromium-review.googlesource.com/1180353
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17288}