[third_party_packages_ng] Add initial API and documentation.
The goal of this recipe+module is to provide a well-documented and
usage-agnostic recipe for building third party packages. Ultimately we
should need to make very few changes to this recipe and module, and all
package definition information will be provided in a subfolder of the
infra.git repo.

R=dnj@chromium.org, hinoka@chromium.org, tandrii@chromium.org

Bug: 877775
Change-Id: Ibb3572cf967911f6837017e729eaeb14e8ba400c
Reviewed-on: https://chromium-review.googlesource.com/1195033
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17293}