[Findit] Flake Detection - Flaky Unit Test Workaround
Some unit tests are flaky depending on thread execution mode, possibly due
to the testing environment's datastore not being cleaned up properly between
tests within the same suite.

This CL is a simple workaround for some of the offending tests to write
different entities to the test datastore to avoid conflicts. The others in
the linked bug are not addressed.

Change-Id: I5a2ffc4250310fd2b86ed9d884cc470a3fbb2451
Bug: 878926
Reviewed-on: https://chromium-review.googlesource.com/1195943
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17295}