[Monorail] Fix autocomplete filtering when there are over 2500.
We used to cap the number of options considered after the first
character was typed, and then only narrow that list down.  That causes
a blind-spot for options that matched the characters typed but that
were over the 2500 limit on options considered.  That meant that even
as the user continues to type, a valid but missing option will never
show up.  E.g., we have tons of options that match "ch" because of
@chromium.org addresses, but a user who's username starts with "ch"
will likely be excluded due to the cap, and then never offered.

We don't need that narrowing-only behavior.  That's an optimization
from before 2005, when JS engines were not as fast as they are today.
I just took it out.

R=ehmaldonado@chromium.org

Bug: monorail:4196
Change-Id: I2b19699de2252965e042325a49c9991cdc850724
Reviewed-on: https://chromium-review.googlesource.com/1198042
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17334}