[Findit] Separate out logic to get flakes with enough occurrences
This is a refactoring CL to seperate out logic in flake util to get
flakes with enough occurrences to report, and the main reason is that
there are flakes that will be reported to Flake Analyzer but not
Monorail.

Once this CL lands, will have a follow-up CL to add logic to report
detected flakes to Flake Analyzer.

Bug: 879683
Change-Id: Idc3e69257bac05c383d99fabab8f0a655c8ca462
Reviewed-on: https://chromium-review.googlesource.com/1199453
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17362}