[Findit] Flake Analzyer - Fixing timeout for swarming tasks
A few flake analyses fail midway through due to too many failed swarming
task attempts analyzing a single data point. The root cause is two fold:

1. Fix the parenthesis error when getting the cushion multiplier set in
   config (e.g. 2x the estimated time it takes to run n iterations). Due
   to the parenthesis error, the actual calculated value was being returned
   as the default value for the cushion in the .get() call, and not being
   in the final multiplied result.
2. Fix the config format that reads where the cushion multipler actually is
   along with unit tests so this doesn't slip past the radar again.

Because of the change to the default config for test cases (now to be in the
same format as what is in production), many other unit tests were updated
as well in unrelated parts of the code.

Bug: 878075
Change-Id: I0ae6f491636e0854399134ae040067c19553ec17
Reviewed-on: https://chromium-review.googlesource.com/1205551
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17404}