[Findit] Flake Analyzer - Surface failed build instance to bug
A user has requested an instance of a failed build be surfaced to the
bug due to the link to the flake occurrence being undiscoverable in the
result UI.

This is a temporary solution, as it still relies on the buildbot build
number to generate the link, which is to be deprecreated. The switch to
build id will be in a follow up, but that change should be invisible to
this one.

Bug: 880571
Change-Id: I4201a93d37084c749b81b517e32ae993701bc176
Reviewed-on: https://chromium-review.googlesource.com/1208022
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17482}