[Findit] Fix bugs in the flaky test sql query
This CL makes the following changes to flaky test sql query:
1. Normalizing step name of a build step by only taking those before
(with patch)|(without patch)|(retry summary), and this is to match the
step name of a test in the test-results table.

2. The returned step name of a flaky test should be the original step
name of the build step instead of the step name of a test in the
test-results table, this is because the step name should be used to
fetch step_metadata through logdog.

Change-Id: I72767de7a38ae0ea4cda9601d4d2c26f3ad02b05
Reviewed-on: https://chromium-review.googlesource.com/1216745
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17493}