[Monorail] Redo permission denied page.
The previous code did not give the user a very helpful error message
because it said that either the item was not found or the user lacked
permission to view it.  And, it only worked for certain pages.  This
restores the earlier working version of the permission denied page
and handles the project not found case for any /p/... URL.

R=jeffcarp@chromium.org

Bug: monorail:4031
Change-Id: Ia52f2a2ba73d5451582c80864ace988a5c7d6bb2
Reviewed-on: https://chromium-review.googlesource.com/1217467
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17494}