[Monorail] Avoid huge phase names that are rejected by trace_service.
This is a follow-up to
https://chromium-review.googlesource.com/c/infra/infra/+/1213946/3/appengine/monorail/businesslogic/work_env.py

We were using a list of project IDs in the name of a profiler phase,
but instead of a list of ints it was actually a dictionary that
had a very long string representation that was rejected by
trace_service.  This was not noticable when running locally because
only GAE-hosted app requests set the X-Cloud-Trace-Context header
that triggers reporting to trace_service.

R=ehmaldonado@chromium.org

Change-Id: I68603a0e9b8d617d0545ab77a5979a0c788e499b
Reviewed-on: https://chromium-review.googlesource.com/1220599
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17513}