[Findit] Fix a bug in Flake Analyzer
CL: https://chromium-review.googlesource.com/c/infra/infra/+/1205551
changed the signature of UnderDailyLimit, but the caller is not updated
accordingly, and due to that UnderDailyLimit is mocked in the tests and
that tests don't assert what arguments UnderDailyLimit is called with,
so this bug was not caught by any tests.

This CL fixes and bug and update the tests.

Change-Id: I1efc0cb2d4c8b21e667bb032c51089d53c4c355c
Reviewed-on: https://chromium-review.googlesource.com/1222767
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17549}