[tricium] Teach Tricium to understand JSON configs
This is particularly helpful for recipe-based analyzers that expect list
parameters (e.g. Fuchsia's format recipe analyzer).

This adds a value_j field to the Selection Config proto, and makes the
value a oneof field to select whether or not it's intended to be JSON.

Change-Id: If3475c0ce3dd82a32c35c1d15042b4efd33f22c5
Reviewed-on: https://chromium-review.googlesource.com/1226637
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Julie Hockett <juliehockett@google.com>
Cr-Commit-Position: refs/heads/master@{#17606}