[buildbucket] Implement search by project
Add support for searching for builds by project.

Given that we plan to change project identity from bucket name to (project id,
bucket name), do not support searching by project and bucket name at the same
time.

This reduces the problem to not accepting a list of buckets, but when listing
the buckets that user has access to, limit them to the ones in the specified
project.

R=tandrii@chromium.org

Bug: 876399
Change-Id: I1f2057e7d840806a76495a3a89ff759b512bc23e
Reviewed-on: https://chromium-review.googlesource.com/1226123
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17616}