[Findit] Flake Analyzer - Surface representative swarming task link to bugs
Developers would like easier access to the stdout of a flaky test, accessible
via the bugs logged. This change surfaces a representative swarming task link
to the bugs. Eventually, we would like to surface the actual stdout itself to
the bug, which is a more involved change.

Bug: 860221
Change-Id: I4353140e0dc89b169f57608fe0413cdb6496489c
Reviewed-on: https://chromium-review.googlesource.com/1226119
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17627}