build_from_tarball: Pass --without-fuchsia to clang's update.py.
This recipe needs the same fix as publish_tarball, which was updated in
https://chromium-review.googlesource.com/1224250.

https://chromium-review.googlesource.com/1211879 added --without-fuchsia new
option to update.py, and it defaults to false just like --without-android.

This caused the "Build From Tarball" bot to start failing in the "Build
clang" step:
https://ci.chromium.org/buildbot/chromium.infra.cron/Build%20From%20Tarball/490

Add a version check and pass --without-fuchsia if we're building a version
that includes it.

Bug: 724204
Change-Id: I5052083675cd59003aaa3227565a9e38c0d08268
Reviewed-on: https://chromium-review.googlesource.com/1227994
Reviewed-by: Michael Moss <mmoss@chromium.org>
Commit-Queue: Raphael Kubo da Costa (CET) <raphael.kubo.da.costa@intel.com>
Cr-Commit-Position: refs/heads/master@{#17635}