[Monorail] Re-land: replace account dropdown with Polymer dropdown
This is re-land of:
https://chromium-review.googlesource.com/c/infra/infra/+/1180353

It was brought to our attention that the dropdown menus in the Monorail
header are not accessible: https://crbug.com/monorail/4132. This replaces
the account header with a Polymer element. If this is successful, we'll
replace the other dropdown menus in the header with Polymer elements.

Eventually the FLT header and Monorail header should merge into one component.

Change-Id: I310d54020c39e00960081652f5472cbf93931565
Reviewed-on: https://chromium-review.googlesource.com/1195963
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17667}