[Findit] Flake Detection UI: add normalized_step_name and test_label_name to order_properties
The paging for ranked flakes dashboard still had issues because the last_occurred_time of flakes are still likely to be the same.

Add normalized_step_name and test_label_name to order_properties as well to ensure the entities can be ordered in a unique way.

Bug: 881620
Change-Id: Ida067dab99f5ee961712bdda957a0f74dac1b47b
Reviewed-on: https://chromium-review.googlesource.com/1234838
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17748}