[cr-buildbucket] clarify public and private symbols in swarming/
I need to do significant refactoring of the way builder's dimensions are
processed and it's currently hard to know what are internal functions and
what are public symbols. Use "_" prefix to make this clear.

Regroup code in flatten_swarmingcfg but not the other ones as this would have
been more disruptive.

I used:
  grep -R -h -o '\bflatten_swarmingcfg\.[^(]\+(' \
      $(git ls-files "*.py" | grep -v _test) | sort | uniq
to determine which symbols are used in other files.

Bug: 880550
Change-Id: Ib2347e8edfd2fa7c4aa30102bdc8651113689aa1
Reviewed-on: https://chromium-review.googlesource.com/1243585
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17863}